requirement.txt
streamlit
tensorflow-macos
tensorflow-metal
numpy
Pillow